add files from repository to file lists
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@18347 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/COPage/classes/class.ilPageObject.php
@@ -1845,7 +1845,10 @@ if ($_GET["pgEdMediaMode"] != "") {echo "ilPageObject::error media"; exit;}
 					$usages = $file_obj->getUsages();
 					if (count($usages) == 0)	// delete, if no usage exists
 					{
-						$file_obj->delete();
+						if ($file_obj->getMode() == "filelist")		// non-repository object
+						{
+							$file_obj->delete();
+						}
 					}
 				}
 			}
